+++ cogs/pause.py
@@ -5,9 +5,8 @@ import datetime
 from utils.colors import BASE_COLOR
 from utils.errors import NoPlayerFound
 from utils import help_utils
-from utils.base_utils import djRole_check
+from utils.base_utils import djRole_check, quiz_check
 from utils import logger
-
 
 @logger.LoggerApplication
 class PlayPauseCommands(commands.Cog):
@@ -18,6 +17,7 @@ class PlayPauseCommands(commands.Cog):
     @app_commands.command(name="pause", description="Pauses current playing track")
     async def pause_command(self, interaction: discord.Interaction):
         if not await djRole_check(interaction, self.logger): return
+        if not await quiz_check(self.bot, interaction, self.logger): return
         try:
             if (player := self.bot.node.get_player(interaction.guild.id)) is None:
                 raise NoPlayerFound("There is no player connected in this guild")
@@ -50,6 +50,7 @@ class PlayPauseCommands(commands.Cog):
     @app_commands.command(name="resume", description="Resumes paused playback")
     async def resume_command(self, interaction: discord.Interaction):
         if not await djRole_check(interaction, self.logger): return
+        if not await quiz_check(self.bot, interaction, self.logger): return
         try:
             if (player := self.bot.node.get_player(interaction.guild.id)) is None:
                 raise NoPlayerFound("There is no player connected in this guild")