+++ cogs/seeking.py
@@ -8,7 +8,7 @@ from utils import help_utils
 from utils.colors import BASE_COLOR
 from utils.regexes import URL_REGEX
 from utils.errors import NoPlayerFound
-from utils.base_utils import convert_to_double, double_to_int, djRole_check
+from utils.base_utils import convert_to_double, double_to_int, djRole_check, quiz_check
 from utils import logger
 from utils.configuration import ConfigurationHandler
 
@@ -21,6 +21,7 @@ class SeekAndRestartCog(commands.Cog):
     @app_commands.command(name="restart", description="Restart current playing track (similiar to seek position: 0:00)")
     async def restart_command(self, interaction: discord.Interaction):
         if not await djRole_check(interaction, self.logger): return
+        if not await quiz_check(self.bot, interaction, self.logger): return
         try:
             if (player := self.bot.node.get_player(interaction.guild.id)) is None:
                 raise NoPlayerFound("There is no player connected in this guild")
@@ -53,6 +54,7 @@ class SeekAndRestartCog(commands.Cog):
     @app_commands.describe(position="Position you want for player to seek ([h:]m:s). If none is provided it will seek forward by 15s")
     async def seek_command(self, interaction: discord.Interaction, position: str=None):
         if not await djRole_check(interaction, self.logger): return
+        if not await quiz_check(self.bot, interaction, self.logger): return
         cfg = ConfigurationHandler(interaction.user.id, user=True)
         try:
             if (player := self.bot.node.get_player(interaction.guild.id)) is None: